stack-trace: First draft of new stack trace code.
I kept the structure of Anton's code, but made the necessary
changes for handling module symbol resolution.

I also took the liberty of eliminating some Xen-specific stuff.
For example, Linux doesn't do the fancy frame pointer inversion
thing that Xen does for exception stacks.

Also, I simplified our regs struct definition, again since the
other one is Xen specific with extras we won't use.

This code assumes that whatever stack we're running on is
LCD_STACK_SIZE bytes and is LCD_STACK_SIZE aligned. This is
true of course for the boot stack, but we'll need to make sure
it remains true for async if we want to debug (I think async uses
larger stacks right now than what our boot stack is).

+++ lcd-domains/include/lcd_domains/microkernel.h
@@ -831,7 +831,8 @@ int __lcd_set_struct_module_hva(struct lcd *caller, cptr_t lcd,
  *
  * Sister function: kernel/kallsyms.c:__sprint_symbol.
  */
-int __lcd_sprint_symbol(char *buffer, hva_t hva, struct module *extra_module);
+unsigned long
+__lcd_sprint_symbol(char *buffer, hva_t hva, struct module *extra_module);
 
 /*
  * --------------------------------------------------
